fix(deps): Switch from dbus-python to pydbus
dbus-python does not install any *.egg-info in site-packages, so running
Mopidy with Mopidy-Notify enabled on Arch fails with

> Disabled extension notify: Dependency The 'dbus-python>=1.2.16,~=1.2' distribution was not found and is required by the application

So we simply replace it with pydbus, which does not seem to have this
kind of problem.

For reference, there [1] is a thread on the python-de mailing list (in
german) on this exact problem.

[1] https://mail.python.org/pipermail/python-de/2017q1/004636.html